Make knobs work

+++ parse_graphics/src/ast.rs
@@ -51,7 +51,7 @@ impl FromStr for Axis {
 #[derive(Clone, Debug)]
 pub enum AstNode {
     Float(f64),
-    Whole(u32),
+    Whole(usize),
     Ident(String),
     Str(String),
     Axis(Axis),
@@ -115,7 +115,7 @@ fn node_from_statement(pair: Pair<Rule>) -> AstNode {
         Rule::vary => AstNode::MdlCommand { command: ParseCommand::Vary, args: get_args(pair) },
         // Primitives
         Rule::float => AstNode::Float(pair.as_str().parse::<f64>().unwrap()),
-        Rule::whole => AstNode::Whole(pair.as_str().parse::<u32>().unwrap()),
+        Rule::whole => AstNode::Whole(pair.as_str().parse::<usize>().unwrap()),
         Rule::axis => AstNode::Axis(pair.as_str().parse::<Axis>().unwrap()),
         Rule::ident => AstNode::Ident(pair.as_str().to_owned()),
         Rule::string => AstNode::Str(pair.as_str().to_owned()),
